Training data for building the predictive model.
This data-set has 1460 records of houses in Ames, a City in Iowa. The data-set includes 80 different predictors / features, based on which the Sale Price needs to be predicted. Some of the features might not be relevant to out task of predicting the Sale Price and hence, will be dropped. At the same time, it is worth to note that we might have to take a deep dive and build / engineer some new features based on the available information. All the decision regarding the dropping or addition of features, are based on exploratory data analysis.